+++ src/contact.11ty.js
@@ -1,12 +1,13 @@
-import pageTypes, {getGeneratedPageData} from './_data/page-types.js'
+import contact from './_data/pages/contact.js'
+import {getGeneratedPageData} from './_lib/page-data.js'
 import {renderContactPage} from './_lib/contact-page.js'
 
 const locale = 'en'
-const pageContent = pageTypes.contact[locale]
+const pageContent = contact[locale]
 
 export default class {
     data() {
-        return getGeneratedPageData(pageTypes.contact, locale, {
+        return getGeneratedPageData(contact, locale, {
             sectionNav: pageContent.sectionNav,
         })
     }