+++ src/fr/dependencies.11ty.js
@@ -1,12 +1,13 @@
 import { legalDocs, renderLegalSection } from '../_lib/legal-docs.js'
-import pageTypes, { getGeneratedPageData } from '../_data/page-types.js'
+import legal from '../_data/pages/legal.js'
+import { getGeneratedPageData } from '../_lib/page-data.js'
 
 const locale = 'fr'
-const pageContent = pageTypes.legal.dependencies[locale]
+const pageContent = legal.dependencies[locale]
 
 export default class {
     data() {
-        return getGeneratedPageData(pageTypes.legal.dependencies, locale, {
+        return getGeneratedPageData(legal.dependencies, locale, {
             sectionNav:legalDocs.dependencies.sectionNav,
         })
     }