+++ src/fr/registration.11ty.js
@@ -1,12 +1,13 @@
-import pageTypes, { getGeneratedPageData } from '../_data/page-types.js'
+import registration from '../_data/pages/registration.js'
+import { getGeneratedPageData } from '../_lib/page-data.js'
 import { renderLaunchRegistrationPage } from '../_lib/launch-registration-page.js'
 
 const locale = 'fr'
-const pageContent = pageTypes.launchRegistration[locale]
+const pageContent = registration[locale]
 
 export default class {
     data() {
-        return getGeneratedPageData(pageTypes.launchRegistration, locale, {
+        return getGeneratedPageData(registration, locale, {
             sectionNav: pageContent.sectionNav,
         })
     }