+++ src/index.11tydata.js
@@ -1,11 +1,11 @@
-import pageTypes from './_data/page-types.js'
+import home from './_data/pages/home.js'
 import i18n from './_data/i18n.js'
 
 const getLocale = (data) => i18n.getLocaleFromUrl(data.page?.url)
-const getContent = (data) => pageTypes.home[getLocale(data)] ?? pageTypes.home.en
+const getContent = (data) => home[getLocale(data)] ?? home.en
 
 export default {
-    layout: pageTypes.home.layout,
+    layout: home.layout,
     eleventyComputed: {
         title:       (data) => getContent(data).title,
         description: (data) => getContent(data).description,